requirements.txt
asgiref==3.4.1
Django==3.2.7
django-multiselectfield==0.1.12
pytz==2021.1
sqlparse==0.4.2
